+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.3] - 2020-01-07
+### Added
+- added css variables --button-color-hover, --button-padding and --border-radius to material-button
+- added slot for hidden file input field to material-button
+- added css variables --error-color and --margin to material-textfield
+
 ## [1.2.2] - 2019-08-11
 ### Added
 - added change event to material-checkbox